refactor(util): remove AbstractStartupStep
It is unclear at this point whether we are going to need this class.

An additional branch was created without this commit - in case we need this class again.
